Support 64-bit integers in ScriptList
(openttd r27104) -Fix [FS#6194]: support 64 bits integere in the
ScriptLists as well

+++ src/script/api/script_list.cpp
@@ -19,9 +19,9 @@
  */
 class ScriptListSorter {
 protected:
-	ScriptList *list;           ///< The list that's being sorted.
+	ScriptList *list;       ///< The list that's being sorted.
 	bool has_no_more_items; ///< Whether we have more items to iterate over.
-	int32 item_next;        ///< The next item we will show.
+	int64 item_next;        ///< The next item we will show.
 
 public:
 	/**
@@ -32,7 +32,7 @@ public:
 	/**
 	 * Get the first item of the sorter.
 	 */
-	virtual int32 Begin() = 0;
+	virtual int64 Begin() = 0;
 
 	/**
 	 * Stop iterating a sorter.
@@ -42,7 +42,7 @@ public:
 	/**
 	 * Get the next item of the sorter.
 	 */
-	virtual int32 Next() = 0;
+	virtual int64 Next() = 0;
 
 	/**
 	 * See if the sorter has reached the end.
@@ -89,7 +89,7 @@ public:
 		this->End();
 	}
 
-	int32 Begin()
+	int64 Begin()
 	{
 		if (this->list->buckets.empty()) return 0;
 		this->has_no_more_items = false;
@@ -99,7 +99,7 @@ public:
 		this->bucket_list_iter = this->bucket_list->begin();
 		this->item_next = *this->bucket_list_iter;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -134,11 +134,11 @@ public:
 		this->item_next = *this->bucket_list_iter;
 	}
 
-	int32 Next()
+	int64 Next()
 	{
 		if (this->IsEnd()) return 0;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -178,7 +178,7 @@ public:
 		this->End();
 	}
 
-	int32 Begin()
+	int64 Begin()
 	{
 		if (this->list->buckets.empty()) return 0;
 		this->has_no_more_items = false;
@@ -193,7 +193,7 @@ public:
 		--this->bucket_list_iter;
 		this->item_next = *this->bucket_list_iter;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -231,11 +231,11 @@ public:
 		this->item_next = *this->bucket_list_iter;
 	}
 
-	int32 Next()
+	int64 Next()
 	{
 		if (this->IsEnd()) return 0;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -270,7 +270,7 @@ public:
 		this->End();
 	}
 
-	int32 Begin()
+	int64 Begin()
 	{
 		if (this->list->items.empty()) return 0;
 		this->has_no_more_items = false;
@@ -278,7 +278,7 @@ public:
 		this->item_iter = this->list->items.begin();
 		this->item_next = (*this->item_iter).first;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -301,11 +301,11 @@ public:
 		if (this->item_iter != this->list->items.end()) item_next = (*this->item_iter).first;
 	}
 
-	int32 Next()
+	int64 Next()
 	{
 		if (this->IsEnd()) return 0;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -343,7 +343,7 @@ public:
 		this->End();
 	}
 
-	int32 Begin()
+	int64 Begin()
 	{
 		if (this->list->items.empty()) return 0;
 		this->has_no_more_items = false;
@@ -352,7 +352,7 @@ public:
 		--this->item_iter;
 		this->item_next = (*this->item_iter).first;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -380,11 +380,11 @@ public:
 		if (this->item_iter != this->list->items.end()) item_next = (*this->item_iter).first;
 	}
 
-	int32 Next()
+	int64 Next()
 	{
 		if (this->IsEnd()) return 0;
 
-		int32 item_current = this->item_next;
+		int64 item_current = this->item_next;
 		FindNext();
 		return item_current;
 	}
@@ -418,7 +418,7 @@ ScriptList::~ScriptList()
 	delete this->sorter;
 }
 
-bool ScriptList::HasItem(int32 item)
+bool ScriptList::HasItem(int64 item)
 {
 	return this->items.count(item) == 1;
 }
@@ -432,7 +432,7 @@ void ScriptList::Clear()
 	this->sorter->End();
 }
 
-void ScriptList::AddItem(int32 item, int32 value)
+void ScriptList::AddItem(int64 item, int64 value)
 {
 	this->modifications++;
 
@@ -444,13 +444,13 @@ void ScriptList::AddItem(int32 item, int32 value)
 	this->SetValue(item, value);
 }
 
-void ScriptList::RemoveItem(int32 item)
+void ScriptList::RemoveItem(int64 item)
 {
 	this->modifications++;
 
 	if (!this->HasItem(item)) return;
 
-	int32 value = this->GetValue(item);
+	int64 value = this->GetValue(item);
 
 	this->sorter->Remove(item);
 	this->buckets[value].erase(item);
@@ -458,13 +458,13 @@ void ScriptList::RemoveItem(int32 item)
 	this->items.erase(item);
 }
 
-int32 ScriptList::Begin()
+int64 ScriptList::Begin()
 {
 	this->initialized = true;
 	return this->sorter->Begin();
 }
 
-int32 ScriptList::Next()
+int64 ScriptList::Next()
 {
 	if (this->initialized == false) {
 		DEBUG(script, 0, "Next() is invalid as Begin() is never called");
@@ -492,20 +492,20 @@ int32 ScriptList::Count()
 	return (int32)this->items.size();
 }
 
-int32 ScriptList::GetValue(int32 item)
+int64 ScriptList::GetValue(int64 item)
 {
 	if (!this->HasItem(item)) return 0;
 
 	return this->items[item];
 }
 
-bool ScriptList::SetValue(int32 item, int32 value)
+bool ScriptList::SetValue(int64 item, int64 value)
 {
 	this->modifications++;
 
 	if (!this->HasItem(item)) return false;
 
-	int32 value_old = this->GetValue(item);
+	int64 value_old = this->GetValue(item);
 	if (value_old == value) return true;
 
 	this->sorter->Remove(item);
@@ -571,7 +571,7 @@ void ScriptList::SwapList(ScriptList *list)
 	list->sorter->Retarget(list);
 }
 
-void ScriptList::RemoveAboveValue(int32 value)
+void ScriptList::RemoveAboveValue(int64 value)
 {
 	this->modifications++;
 
@@ -581,7 +581,7 @@ void ScriptList::RemoveAboveValue(int32 value)
 	}
 }
 
-void ScriptList::RemoveBelowValue(int32 value)
+void ScriptList::RemoveBelowValue(int64 value)
 {
 	this->modifications++;
 
@@ -591,7 +591,7 @@ void ScriptList::RemoveBelowValue(int32 value)
 	}
 }
 
-void ScriptList::RemoveBetweenValue(int32 start, int32 end)
+void ScriptList::RemoveBetweenValue(int64 start, int64 end)
 {
 	this->modifications++;
 
@@ -601,7 +601,7 @@ void ScriptList::RemoveBetweenValue(int32 start, int32 end)
 	}
 }
 
-void ScriptList::RemoveValue(int32 value)
+void ScriptList::RemoveValue(int64 value)
 {
 	this->modifications++;
 
@@ -695,7 +695,7 @@ void ScriptList::RemoveList(ScriptList *list)
 	}
 }
 
-void ScriptList::KeepAboveValue(int32 value)
+void ScriptList::KeepAboveValue(int64 value)
 {
 	this->modifications++;
 
@@ -705,7 +705,7 @@ void ScriptList::KeepAboveValue(int32 value)
 	}
 }
 
-void ScriptList::KeepBelowValue(int32 value)
+void ScriptList::KeepBelowValue(int64 value)
 {
 	this->modifications++;
 
@@ -715,7 +715,7 @@ void ScriptList::KeepBelowValue(int32 value)
 	}
 }
 
-void ScriptList::KeepBetweenValue(int32 start, int32 end)
+void ScriptList::KeepBetweenValue(int64 start, int64 end)
 {
 	this->modifications++;
 
@@ -725,7 +725,7 @@ void ScriptList::KeepBetweenValue(int32 start, int32 end)
 	}
 }
 
-void ScriptList::KeepValue(int32 value)
+void ScriptList::KeepValue(int64 value)
 {
 	this->modifications++;
